uploader integration works - demo commit

+++ src/GeoNodePy/geonode/maps/views.py
@@ -878,11 +878,17 @@ def upload_layer(request):
                         )
 
                 if settings.USE_UPLOADER:
-                     saved_layer, import_session = saved_layer
+                     # this is really a import_session object to be clear
+                     import_session = saved_layer
+                     # @todo session objects prolly should be handled more carefully
                      request.session['import_session'] = import_session
+                     for k in form.cleaned_data.keys():
+                        if k.endswith('_file'):
+                            form.cleaned_data.pop(k)
+                     request.session['import_form'] = form.cleaned_data
+                     request.session['import_base_file'] = base_file
                      # only feature types have attributes
                      if hasattr(import_session.tasks[0].items[0].resource,"attributes"):
-                         request.session['final_step'] = saved_layer.get_absolute_url() + "?describe"
                          return HttpResponse(json.dumps({
                          "success": True,
                          "redirect_to": reverse('data_upload2')}))
@@ -896,7 +902,7 @@ def upload_layer(request):
                     "success": False,
                     "errors": ["Unexpected error during upload: " + escape(str(e))]}))
             finally:
-                if tempdir is not None:
+                if tempdir is not None and not settings.USE_UPLOADER:
                     shutil.rmtree(tempdir)
         else:
             errors = []
@@ -914,8 +920,8 @@ def upload_layer2(request):
              RequestContext(request, upload_step2_context(request))
          )
      elif request.method == 'POST':
-         upload_step2(request)
-         return HttpResponseRedirect(request.session['final_step'])
+         saved_layer = upload_step2(request)
+         return HttpResponseRedirect(saved_layer.get_absolute_url() + "?describe")
 
 @login_required
 @csrf_exempt
